+++ lib/src/banner/banner_ad_widget.dart
@@ -16,6 +16,10 @@ const _viewType = 'banner_admob';
 ///   - https://developers.google.com/admob/android/banner on Android
 ///   - https://developers.google.com/admob/ios/banner on iOS
 class BannerAd extends StatefulWidget {
+  /// Creates a new Banner Ad.
+  /// `size` can NOT be null. If so, an `AssertionError` is thrown
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-banner-ad)
   const BannerAd({
     Key key,
     this.builder,
@@ -23,7 +27,8 @@ class BannerAd extends StatefulWidget {
     this.size = BannerSize.ADAPTIVE,
     this.error,
     this.loading,
-  })  : assert(size != null),
+    this.unitId,
+  })  : assert(size != null, 'A size must be set'),
         super(key: key);
 
   /// The builder of the ad. The ad won't be reloaded if this changes
@@ -98,7 +103,8 @@ class BannerAd extends StatefulWidget {
   /// For more info, visit the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Using-the-controller-and-listening-to-banner-events)
   final BannerAdController controller;
 
-  /// The size of the Ad. `BannerSize.ADAPTIVE` is the default
+  /// The size of the Ad. `BannerSize.ADAPTIVE` is the default.
+  /// This can NOT be null. If so, throws an `AssertionError`
   ///
   /// ## Sizes
   ///
@@ -136,6 +142,11 @@ class BannerAd extends StatefulWidget {
   /// For more info, visit the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-banner-ad#creating-an-ad)
   final BannerSize size;
 
+  /// The unitId used by this `BannerAd`.
+  /// If changed after loaded it'll be reloaded with the new ad unit id.\
+  /// If null, defaults to `MobileAds.bannerAdUnitId`
+  final String unitId;
+
   @override
   _BannerAdState createState() => _BannerAdState();
 }
@@ -157,16 +168,20 @@ class _BannerAdState extends State<BannerAd>
   // @override
   // void didUpdateWidget(BannerAd oldWidget) {
   //   super.didUpdateWidget(oldWidget);
-  //   if (oldWidget.controller == null && widget.controller != null) {
-  //     attachNewController();
-  //     controller.changeController(controller.id);
-  //     controller.load();
-  //   }
+  //   // if ((oldWidget.unitId == null && widget.unitId != null) ||
+  //   //     (oldWidget.unitId != widget.unitId)) {
+  //   //   controller.load();
+  //   // }
+  //   // if (oldWidget.controller == null && widget.controller != null) {
+  //   //   attachNewController();
+  //   //   controller.changeController(controller.id);
+  //   //   controller.load();
+  //   // }
   // }
 
   void attachNewController() {
     controller = widget.controller ?? BannerAdController();
-    controller.attach();
+    controller.attach(true);
     controller.onEvent.listen((e) {
       final event = e.keys.first;
       final info = e.values.first;
@@ -177,9 +192,6 @@ class _BannerAdState extends State<BannerAd>
           height = (info as int)?.toDouble();
           setState(() => state = event);
           break;
-        case BannerAdEvent.undefined:
-          setState(() {});
-          break;
         default:
           break;
       }
@@ -188,7 +200,12 @@ class _BannerAdState extends State<BannerAd>
 
   @override
   void dispose() {
-    controller?.dispose();
+    // dispose the controller only if the controller was
+    // created by the ad.
+    if (widget.controller == null)
+      controller?.dispose();
+    else
+      controller?.attach(false);
     super.dispose();
   }
 
@@ -210,7 +227,7 @@ class _BannerAdState extends State<BannerAd>
         final params = <String, dynamic>{};
         params.addAll({
           'controllerId': controller.id,
-          'unitId': MobileAds.bannerAdUnitId,
+          'unitId': widget.unitId ?? MobileAds.bannerAdUnitId,
           'size_height': height ?? -1,
           'size_width': width,
         });