+++ lib/src/native/gradient/radial.dart
@@ -1,13 +1,12 @@
 part of 'gradient.dart';
 
 class AdRadialGradient extends AdGradient {
-
   /// Creates a RadialGradient
-  /// 
+  ///
   /// Uses a [GradientDrawable](https://developer.android.com/reference/android/graphics/drawable/GradientDrawable)
   /// on Android
-  /// 
-  /// - You must specify at least two [colors]. 
+  ///
+  /// - You must specify at least two [colors].
   /// - The default value for [radius] is 1000
   /// - The default value for [center] is `Alignment(0.5, 0.5)`
   AdRadialGradient({